+++ example/build.js
@@ -5,7 +5,7 @@ const args = process.argv.slice(2);
 const files = args.map(arg => path.resolve(process.cwd(), arg));
 
 process.stdout.write("Running tests: ");
-const result = await TestRunner.create().runInChildProcessAsync(files, { notifyFn: reportProgress });
+const result = await TestRunner.create().runInChildProcessAsync(files, { onTestCaseResult: reportProgress });
 console.log("\n" + result.render("\n") + "\n");
 
 function reportProgress(testCase) {